Changing return type of `__iter__` to `Iterator`.
The previous return type of `Iterable` is incorrect as can be seen in
this example:
https://docs.python.org/3/library/typing.html#nominal-vs-structural-subtyping

Also this documentation mentions that `__iter__` returns the iterator
object itself:
https://docs.python.org/3/library/stdtypes.html#iterator.__iter__

The incorrect `Iterable` type causes numerous of error when analyzed
by a static type checker such as `mypy`. The switch to `Iterator`
solves these issues.